# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "中医药知识库 API 服务 —— 方剂、单味药、对药、论文、分子信息、名家经验管理",
        "title": "AI Medical Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/compounds": {
            "get": {
                "description": "分页查询化合物，支持关键词搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compounds"
                ],
                "summary": "获取化合物列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/medicalagent_internal_model.MolecularInfo"
                                    }
                                },
                                "page": {
                                    "type": "integer"
                                },
                                "page_size": {
                                    "type": "integer"
                                },
                                "total": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员新增化合物记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compounds"
                ],
                "summary": "创建化合物",
                "parameters": [
                    {
                        "description": "化合物信息",
                        "name": "compound",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medicalagent_internal_model.MolecularInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_model.MolecularInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/compounds/{id}": {
            "get": {
                "description": "根据 Record Number 获取化合物详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compounds"
                ],
                "summary": "获取化合物详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record Number",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_model.MolecularInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员更新现有化合物记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compounds"
                ],
                "summary": "更新化合物",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record Number",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "化合物信息",
                        "name": "compound",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medicalagent_internal_model.MolecularInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员删除指定化合物记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compounds"
                ],
                "summary": "删除化合物",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record Number",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "medicalagent_internal_model.MolecularInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fda_pharmacology_summary": {
                    "type": "string"
                },
                "heavy_atom_count": {
                    "type": "integer"
                },
                "hydrogen_bond_acceptor_count": {
                    "type": "integer"
                },
                "hydrogen_bond_donor_count": {
                    "type": "integer"
                },
                "inchi": {
                    "type": "string"
                },
                "inchi_key": {
                    "type": "string"
                },
                "livertox_summary": {
                    "type": "string"
                },
                "molecular_formula": {
                    "type": "string"
                },
                "molecular_weight": {
                    "type": "number"
                },
                "record_description": {
                    "type": "string"
                },
                "record_number": {
                    "type": "integer"
                },
                "record_title": {
                    "type": "string"
                },
                "rotatable_bond_count": {
                    "type": "integer"
                },
                "smiles": {
                    "type": "string"
                },
                "xlogp3": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgentTokenAuth": {
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}